Data extraction from terminal for test case of prgm 1

+++ Main.py
@@ -6,6 +6,9 @@ import sys
 import renormalize as renorm
 
 def function(option):
+
+    UI = False  #FIXME
+
 #Exports data requested by the user into text files (necessary to generate plots)
     userInput = ided.datExp(option,True,True)
 
@@ -20,30 +23,30 @@ def function(option):
 #Makes plot (.012 s)
     ided.pltFileExp(option,userInput[6],userInput[4],userInput[0],userInput[1],userInput[2],userInput[7],userInput[3],True)
 
-
-#This code creates the .git file which is the actual plot
-    os.chdir("Output/gnuPlot")
-    directory = os.getcwd()
-    try:
-        newest = max(glob.iglob(directory+"/*.plt"),key=os.path.getctime)
-        newest = newest.replace(os.getcwd()+"/","")
-        os.system("gnuplot "+newest)
-    except:
-        print('No new plot')
-    
+    if UI:
+    #This code creates the .git file which is the actual plot
+        os.chdir("Output/gnuPlot")
+        directory = os.getcwd()
+        try:
+            newest = max(glob.iglob(directory+"/*.plt"),key=os.path.getctime)
+            newest = newest.replace(os.getcwd()+"/","")
+            os.system("gnuplot "+newest)
+        except:
+            print('No new plot')
 #This code puts restarts the program so it can be used again
-    os.chdir("..")
-    os.chdir("..")
-    os.system("python3 Main.py "+option)
-    newest = "Output/gnuPlot/"+newest.replace(".plt",".gif")
-    if os.path.isfile(newest):
-        os.system("rm "+newest)
-    try:
-        os.system("mv Output/gnuPlot/*.dat Output/gnuPlot/OutputData")
-        os.system("mv Output/gnuPlot/*.plt Output/gnuPlot/OutputData")
-    except:
+        os.chdir("..")
+        os.chdir("..")
+        os.system("python3 Main.py "+option)
+        newest = "Output/gnuPlot/"+newest.replace(".plt",".gif")
+        if os.path.isfile(newest):
+            os.system("rm "+newest)
+        try:
+            os.system("mv Output/gnuPlot/*.dat Output/gnuPlot/OutputData")
+            os.system("mv Output/gnuPlot/*.plt Output/gnuPlot/OutputData")
+        except:
+            pass
+    else:
         pass
-
 
 option = sys.argv[-1]
 function(option)
